tiffile warning filter has to be on top level to make effect
